# backend/hhrr_app/static/swagger/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"description": "This API provides the Plena Global Holdings software development team with the necessary knowledge base for the integration of the different applications with the services offered by the human resources department.",
		"version": "1.0.0",
		"title": "HUMMAN RESOURCES DEPARTMENT API",
		"license": {
			"name": "MIT",
			"url": "https://opensource.org/licenses/MIT"
		}
	},
	"servers": [
		{
			"url": "/api/v1/"
		}
	],
	"components": {
		"schemas": {
...
			"tbl_family_nucleus": {
				"type": "object",
				"properties": {
					"ccn_family_nucleus": {
						"type": "integer",
						"example": 1,
						"description": "Incremental integer number assigned by the system automatically when recording a new record."
					},
					"ccn_employee": {
						"type": "integer",
						"minimum": 1,
						"maximum": 10000,
						"example": 1,
						"description": "Foreign key with table tbl_employee"
					},
					"ccn_marital_status": {
						"type": "number",
						"minimum": 1,
						"maximum": "Jose Ruiz",
						"example": 1,
						"description": "Foreign key with table tbl_martial_status"
					},
					"number_of_children": {
						"type": "number",
						"minimum": 1,
						"maximum": 10000,
						"example": 1,
						"description": "Quantity of childrens"
					},
					"ccn_type_id": {
						"type": "number",
						"minimum": 1,
						"maximum": 10000,
						"example": 1,
						"description": "Type of document of child"
					},
					"number_id": {
						"type": "number",
						"minimum": 1,
						"maximum": 10000,
						"example": 98056369,
						"description": "Number ID from child"
					},
					"ccn_auto_perceived_gender": {
						"type": "number",
						"minimum": 1,
						"maximum": 10000,
						"example": 1,
						"description": "Foreign key with table tbl_auto_perceived_gender"
					},
					"first_name": {
						"type": "string",
						"minimum": 1,
						"maximum": 10000,
						"example": "Juan",
						"description": "First name from child"
					},
					"middle_name": {
						"type": "string",
						"minimum": 1,
						"maximum": 10000,
						"example": "Andres",
						"description": "Middle name from child"
					},
					"first_last_name": {
						"type": "string",
						"minimum": 1,
						"maximum": 10000,
						"example": "Rincon",
						"description": "First last name from child"
					},
					"second_last_name": {
						"type": "string",
						"minimum": 1,
						"maximum": 10000,
						"example": "Guevara",
						"description": "Scecond last name from child"
					},
					"date_of_birth": {
						"type": "date",
						"minimum": 1,
						"maximum": 10000,
						"example": "1990-04-14",
						"description": "Date of birth"
					},
					"age": {
						"type": "number",
						"minimum": 1,
						"maximum": 10000,
						"example": 5,
						"description": "Age from child"
					},
					"age_range": {
						"type": "string",
						"minimum": 1,
						"maximum": 10000,
						"example": "20-24",
						"description": "Age range from child"
					},
					"ccn_schooling_level": {
						"type": "number",
						"minimum": 1,
						"maximum": 10000,
						"example": 1,
						"description": "Foreign key with table tbl_schooling_level"
					}
				}
			},
...
			"tbl_department": {
				"type": "object",
				"properties": {
					"ccn_department": {
						"type": "integer (pk) auto_increment",
						"example": 1,
						"description": "Incremental integer number assigned by the system automatically when recording a new record."
					},
					"id_department": {
						"type": "integer",
						"minimum": 0,
						"maximum": 10000,
						"example": 11,
						"description": "DIAN idintification number"
					},
					"descripcion_department": {
						"type": "string",
						"example": "CUNDINAMARCA",
						"description": "department name from employee"
					},
					"ccn_country": {
						"type": "integer",
						"minimum": 0,
						"maximum": 40,
						"example": 1,
						"description": "Foreign key with the table tbl_countrys"
					}
				}
			},
...
		}
	},
	"paths": {
...
		"/family_nucleus": {
			"post": {
				"tags": [
					"Basic data from family nucleus"
				],
				"description": "Create a family nucleus details data to human resource platform",
				"requestBody": {
					"description": "Post family nucleus details data to database",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/tbl_family_nucleus"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "CREATED",
						"schema": {
							"$ref": "#/components/schemas/tbl_family_nucleus"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_family_nucleus"
						}
					}
				}
			},
			"get": {
				"tags": [
					"Basic data from family nucleus"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_family_nucleus"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_family_nucleus"
						}
					}
				}
			}
		},
		"/family_nucleus/{ccn_family_nucleus}": {
			"get": {
				"tags": [
					"Basic data from family nucleus"
				],
				"parameters": [
					{
						"name": "ccn_family_nucleus",
						"in": "path",
						"description": "todo id to update",
						"required": true,
						"type": "integer"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_family_nucleus"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_family_nucleus"
						}
					}
				}
			},
			"put": {
				"tags": [
					"Basic data from family nucleus"
				],
				"parameters": [
					{
						"name": "ccn_family_nucleus",
						"in": "path",
						"description": "todo id to update",
						"required": true,
						"type": "integer"
					}
				],
				"requestBody": {
					"description": "Update Existing todo On Database",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/tbl_family_nucleus"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_family_nucleus"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_family_nucleus"
						}
					}
				}
			},
			"delete": {
				"tags": [
					"Basic data from family nucleus"
				],
				"parameters": [
					{
						"name": "ccn_family_nucleus",
						"in": "path",
						"description": "product id to delete",
						"required": true,
						"type": "integer"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_family_nucleus"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_family_nucleus"
						}
					}
				}
			}
		},
...
		"/department": {
			"post": {
				"tags": [
					"Basic data from Department"
				],
				"description": "Create a department data to human resource platform",
				"requestBody": {
					"description": "Post department data to database",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/tbl_department"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "CREATED",
						"schema": {
							"$ref": "#/components/schemas/tbl_department"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_department"
						}
					}
				}
			},
			"get": {
				"tags": [
					"Basic data from Department"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_department"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_department"
						}
					}
				}
			}
		},
		"/department/{ccn_department}": {
			"get": {
				"tags": [
					"Basic data from Department"
				],
				"parameters": [
					{
						"name": "ccn_department",
						"in": "path",
						"description": "todo id to update",
						"required": true,
						"type": "integer"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_department"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_department"
						}
					}
				}
			},
			"put": {
				"tags": [
					"Basic data from Department"
				],
				"parameters": [
					{
						"name": "ccn_department",
						"in": "path",
						"description": "todo id to update",
						"required": true,
						"type": "integer"
					}
				],
				"requestBody": {
					"description": "Update Existing todo On Database",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/tbl_department"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_department"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_department"
						}
					}
				}
			},
			"delete": {
				"tags": [
					"Basic data from Department"
				],
				"parameters": [
					{
						"name": "ccn_department",
						"in": "path",
						"description": "product id to delete",
						"required": true,
						"type": "integer"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_department"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_department"
						}
					}
				}
			}
		},
...
	}
}